requirements.txt
# requirements.txt
# Python dependencies for Gait Detection System
# Updated for compatibility with PyTorch 2.x and Python 3.8+

# ==================== Core Deep Learning ====================
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0

# ==================== Data Processing ====================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scipy>=1.10.0,<1.15.0

# ==================== Machine Learning & Metrics ====================
scikit-learn>=1.3.0,<1.6.0

# ==================== Visualization ====================
matplotlib>=3.7.0,<3.10.0
seaborn>=0.12.0,<0.14.0
plotly>=5.18.0,<6.0.0  # Interactive plots for dashboard

# ==================== Explainability ====================
shap>=0.44.0,<0.46.0
lime>=0.2.0.1  # LIME explainability for all models

# ==================== Configuration ====================
pyyaml>=6.0,<7.0

# ==================== Progress Bars ====================
tqdm>=4.65.0,<5.0.0

# ==================== Testing ====================
pytest>=7.4.0,<9.0.0
coverage>=7.2.0,<8.0.0
unittest-xml-reporting>=3.2.0  # For test reports

# ==================== Jupyter Notebooks ====================
jupyter>=1.0.0,<2.0.0
ipykernel>=6.25.0,<7.0.0
notebook>=7.0.0,<8.0.0
jupyterlab>=4.0.0,<5.0.0  # Alternative to notebook

# ==================== Jupyter Extensions ====================
ipympl>=0.9.0,<1.0.0  # For interactive plots
ipywidgets>=8.1.0,<9.0.0  # For interactive widgets

# ==================== Model Export ====================
onnx>=1.15.0,<2.0.0
onnxruntime>=1.16.0,<2.0.0

# ==================== Web Dashboard ====================
streamlit>=1.28.0,<2.0.0  # Web dashboard framework
openpyxl>=3.1.0,<4.0.0  # Excel file support

# ==================== Additional Utilities ====================
pillow>=10.0.0,<11.0.0  # Image processing
h5py>=3.9.0,<4.0.0  # HDF5 file support
tensorboard>=2.14.0,<3.0.0  # Optional: for training visualization

